Add a build cache to speed up builds, because **a blazingly fast library deserves a blazingly fast build:)**

Hi,

While working on Sonic, I noticed that the current build script does not utilize any caching mechanism.
As a result, even a small change can cause a full rebuild from scratch, which is quite time-consuming.

To address this, I added a cache. With the new script, build times are reduced by up to a whopping **96%** (measured on my local machine).

### The old build vs. the blazingly fast cached build 

```
(venv) alex@x sonic_build % ./scripts/bench_x86_cache_build.sh
Compiler      : clang
Old script    : /Users/alex/Desktop/soniker/sonic_build/scripts/build-x86.sh
Cache script  : /Users/alex/Desktop/soniker/sonic_build/scripts/build-x86_cache.sh
Runs per case : 5

----- Old script (/Users/alex/Desktop/soniker/sonic_build/scripts/build-x86.sh) -----
Run 1: 147 s
Run 2: 144 s
Run 3: 144 s
Run 4: 142 s
Run 5: 142 s
Shortest (old): 142 s

----- New script (/Users/alex/Desktop/soniker/sonic_build/scripts/build-x86_cache.sh) -----
Run 1 (with --force): 146 s
Run 2: 6 s
Run 3: 5 s
Run 4: 5 s
Run 5: 6 s
Shortest (new): 5 s

The new script reduced build time by up to 96.47%
```

I did not replace the original script:
- It’s unclear whether a full rebuild is intentional, or if a different build script is used, which is not open sourced yet.
- The C code has been stable so that build script doesn't get run very often. Thus, using the old script may not be such an inconvenience.
- The new script has some limitations: it assumes that asm2asm output files are named `xxx_text_amd64.go` and `xxx_subr.go`, which can’t be deduced from the build script itself.

That said, caching the output of asm2asm is necessary. On my machine, the asm2asm phase takes about 125 seconds, while compiling the C code only about 30 seconds. So caching asm2asm output saves the most time.

### What’s new:
- `scripts/build-x86_cache.sh`: the new build script with caching.
- Helper scripts:
    - `scripts/test_x86_cache_build.sh`: verifies correctness by comparing outputs with the original build.
    - `scripts/bench_x86_cache_build.sh`: benchmarks build time improvements.
    - `scripts/build-x86_cache_output.sh`: lists output files, particularly the suffix of asm2asm outputs, based on the original script.
    - `scripts/build-x86_cache_time.sh`: compares C compilation time vs. asm2asm execution time.

Other file changes are generated as a result of initial script run.

### Main ideas:

The whole build process can be divided into three stages:
- *.tmpl -> *.go
- C -> .s
- asm2asm (*.go + .s -> *_subr.go + *_text_amd64.go)

The output of each stage is cached under a filename derived from the SHA‑1 of all its inputs (content‑addressable storage).
If any upstream input changes, its hash changes, the cache misses, and every downstream step is forced to recompute -- guaranteeing correctness while eliminating redundant work.

`--force` can be used to purge the cache and the script itself doesn't have a mechanism to automatically do so.

### Lastly

If this PR is considered mergeable, helper scripts should be removed after review.
And, please consider merging this into the main branch. After all, **a blazingly fast library deserves a blazingly fast build:)**

Feedback and reviews are most welcome.

最后，向开发、维护 Sonic 库的硬核大佬敬礼～